chore: Add find customer by id

+++ src/modules/customers/services/DeleteCustomerService.ts
@@ -1,6 +1,8 @@
 import 'reflect-metadata';
 import { inject, injectable } from 'tsyringe';
 
+import AppError from '@shared/errors/AppError';
+
 import ICustomersRepository from '../repositories/ICustomersRepository';
 
 @injectable()
@@ -12,6 +14,14 @@ export default class CreateCustomerSession {
 
   // Create a new customer
   public async execute(id: string): Promise<void> {
-    this.customersRepository.deleteCustomerById(id);
+    const checkIfCustomerExist = await this.customersRepository.findCustomerById(
+      id,
+    );
+
+    if (!checkIfCustomerExist) {
+      throw new AppError('Customer does not exists', 404);
+    }
+
+    await this.customersRepository.deleteCustomerById(id);
   }
 }
